# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/business-instagram": {
            "post": {
                "description": "Business Instagramを作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business-instagram"
                ],
                "summary": "Business Instagram作成",
                "parameters": [
                    {
                        "description": "作成データ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.BusinessInstagram"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "作成されたBusiness Instagram",
                        "schema": {
                            "$ref": "#/definitions/res.BusinessInstagram"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/business-instagram/{id}": {
            "put": {
                "description": "Business Instagramを更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "business-instagram"
                ],
                "summary": "Business Instagram更新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Business Instagram ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新データ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.BusinessInstagram"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新されたBusiness Instagram",
                        "schema": {
                            "$ref": "#/definitions/res.BusinessInstagram"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "見つかりません",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/google-business": {
            "get": {
                "description": "Google Businessの一覧を取得します（ページング対応）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "google-business"
                ],
                "summary": "Google Business一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "取得件数（デフォルト: 20）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "オフセット（デフォルト: 0）",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Google Business一覧",
                        "schema": {
                            "$ref": "#/definitions/res.GoogleBusinessList"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/google-business/fetch": {
            "get": {
                "description": "Google Businessを同期します",
                "tags": [
                    "google-business"
                ],
                "summary": "Google Businessの同期",
                "responses": {
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sync": {
            "post": {
                "description": "全ての顧客のデータを同期します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "全顧客データ同期",
                "responses": {
                    "200": {
                        "description": "全顧客同期完了",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/sync/{id}": {
            "post": {
                "description": "全ての顧客のデータを同期します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync"
                ],
                "summary": "全顧客データ同期",
                "responses": {
                    "200": {
                        "description": "全顧客同期完了",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/token": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "トークンを取得します。",
                "responses": {
                    "200": {
                        "description": "トークン情報",
                        "schema": {
                            "$ref": "#/definitions/res.Token"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "トークンを保存します。",
                "responses": {
                    "200": {
                        "description": "更新完了",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/token/check": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token"
                ],
                "summary": "トークンの認証情報を取得する",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/wordpress-instagram": {
            "get": {
                "description": "Wordpress Instagramの一覧を取得します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wordpress-instagram"
                ],
                "summary": "Wordpress Instagram一覧取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "取得件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "オフセット",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名前",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "WordPress URL",
                        "name": "wordpress",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Instagram ID",
                        "name": "instagram_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ステータス",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "削除ハッシュ",
                        "name": "delete_hash",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "顧客タイプ",
                        "name": "customer_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wordpress Instagram一覧",
                        "schema": {
                            "$ref": "#/definitions/res.WordpressInstagramList"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Wordpress Instagramを作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wordpress-instagram"
                ],
                "summary": "Wordpress Instagram作成",
                "parameters": [
                    {
                        "description": "作成データ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CreateWordpressInstagram"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "作成されたWordpress Instagram",
                        "schema": {
                            "$ref": "#/definitions/res.WordpressInstagram"
                        }
                    },
                    "400": {
                        "description": "不正なリクエスト",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部サーバーエラー",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "req.BusinessInstagram": {
            "type": "object",
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "instagram_id": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "req.CreateWordpressInstagram": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "delete_hash": {
                    "type": "boolean"
                },
                "instagram_id": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "wordpress_domain": {
                    "type": "string"
                }
            }
        },
...
        "res.BusinessInstagram": {
            "type": "object",
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instagram_id": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "res.GoogleBusiness": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "res.GoogleBusinessList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "google_business_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/res.GoogleBusiness"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "res.Token": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "res.WordpressInstagram": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "delete_hash": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "instagram_id": {
                    "type": "string"
                },
                "instagram_name": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "wordpress_domain": {
                    "type": "string"
                },
                "wordpress_site_title": {
                    "type": "string"
                }
            }
        },
...
        "res.WordpressInstagramList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "instagram_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/res.WordpressInstagram"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}